fix fav Bug and add language logic

+++ app/src/main/java/ahmed/adel/sleeem/clowyy/souq/ui/fragments/details/adapter/RecommendedDetailsRA.kt
@@ -48,7 +48,7 @@ class RecommendedDetailsRA (val context:Context) : RecyclerView.Adapter<Recommen
                 binding.tvOffPercentage.visibility = View.INVISIBLE
                 binding.tvOldCost.visibility = View.INVISIBLE
             }
-            binding.ratingBar.rating = (product.rating/2.0f)
+            binding.ratingBar.rating = (product.rating/2.0f).toFloat()
 
             if(itemClickListener != null) {
                 setOnClickListener {
