+++ app/src/main/java/ahmed/adel/sleeem/clowyy/souq/ui/fragments/explore/adapter/SearchSucceedAdapter.kt
@@ -59,7 +59,7 @@ class SearchSucceedAdapter (val context:Context) :
                 binding.tvOffPercentage.visibility = View.INVISIBLE
                 binding.tvOldCost.visibility = View.INVISIBLE
             }
-            binding.ratingBar.rating = (product.rating/2.0f)
+            binding.ratingBar.rating = (product.rating/2.0f).toFloat()
 
             if(itemClickListener != null) {
                 setOnClickListener {
